+++ Spike.cpp
@@ -4,7 +4,7 @@ Spike::Spike() : Entity() {
 
 }
 
-Spike::Spike(uint8_t sprite_index, SpikeDirection direction, float x, float y) : Entity(sprite_index, x, y) {
+Spike::Spike(uint16_t sprite_index, SpikeDirection direction, float x, float y) : Entity(sprite_index, x, y) {
 	this->direction = direction;
 }
